[API] node: implement and expose cmark_node_set_type.
This is necessary for extensions implementing vertical rules.

An example is setext headings:

A heading
---------

When cmark parses 'A heading', it first creates a paragraph
block to contain it, it's only when cmark parses the second line
that the type of the block is changed to the
CMARK_NODE_TYPE_HEADING type.