[API]: node: Expose string_content.
By implementing cmark_node_get_string_content and
cmark_node_set_string_content.

This is useful for vertical rules in extensions, as they may need
to access it in order to decide whether to update the block.

Unfortunately, this overlaps with get_literal and set_literal.

As far as I can tell we should deprecate these functions, and
have them follow the get_string_content code path and
set_string_content for the while.